Renaming Bot class to CognitumClient and other minor changes
