+++ lib/classes/CognitumClient.js
@@ -7,13 +7,12 @@ const MessageParser = require("./MessageParser.js");
 const TasksQueue = require("./tasks/TasksQueue");
 const MonitoringStatsTask = require("./tasks/internal/MonitoringStatsTask");
 const { formatTimeString, log } = require("./Utils.js");
-const VoiceStateManager = require("./VoiceStateManager");
-const BaseError = require("../classes/base/BaseError");
+const VoiceStateManager = require("./statistics/VoiceStateManager");
 
 /**
  * Extended class from Discord Client.
  */
-class Bot extends Client {
+class CognitumClient extends Client {
 	/**
 	 * Tasks queue.
 	 * @type {TasksQueue}
@@ -122,4 +121,4 @@ class Bot extends Client {
 	}
 }
 
-module.exports = Bot;
+module.exports = CognitumClient;
