+++ lib/classes/Database.js
@@ -7,8 +7,7 @@ const MessageStatisticsModel = require("../database/models/MessageStatisticsMode
 const UserModel = require("../database/models/UserModel");
 const VoiceStatisticsModel = require("../database/models/VoiceStatisticsModel");
 const TaskModel = require("../database/models/TaskModel");
-
-const { createModuleLog } = require("../classes/Utils.js");
+const { createModuleLog } = require("../classes/Utils");
 const log = createModuleLog("Database");
 const Config = require("../classes/ConfigManager");
