+++ lib/classes/Utils.js
@@ -2,7 +2,7 @@
  * # Utilities
  * Object with useful utilities may be used in the code.
  */
-const Utils = {
+class Utils {
 	/**
 	 * Centralized logging function. Used for unified style of logging.
 	 * @param {Cognitum.LogType} type Type of log.
@@ -11,10 +11,10 @@ const Utils = {
 	 * @todo Different logging colors from config used by type
 	 * @todo Different logging "headers" for every type
 	 */
-	log(type, message) {
+	static log(type, message) {
 		const now = new Date();
 		console.log(`${now.toLocaleString()} [${type}] ${message}`);
-	},
+	}
 
 	/**
 	 * Logging function factory. Uses closures for inserting module name into the original log function.
@@ -28,11 +28,11 @@ const Utils = {
 	 * log('log', 'Loading system module...');
 	 * // 8/7/2020, 12:20:36 AM [SystemModule] [log] Loading system module...
 	 */
-	createModuleLog(moduleName) {
+	static createModuleLog(moduleName) {
 		return (type, message) => {
 			Utils.log(moduleName, `[${type}] ${message}`);
 		};
-	},
+	}
 
 	/**
 	 * Get file name from basename with extension.
@@ -47,11 +47,11 @@ const Utils = {
 	 * Utils.fileName(basenameWithoutExtension);
 	 * // "README"
 	 */
-	fileName(basename) {
+	static fileName(basename) {
 		return basename.split(".")
 			.slice(0, 1 - +basename.includes(".") * 2)
 			.join(".");
-	},
+	}
 
 	/**
 	 * Get file extension from basename.
@@ -66,12 +66,12 @@ const Utils = {
 	 * Utils.fileExtension(basenameWithoutExtension);
 	 * // false
 	 */
-	fileExtension(basename) {
+	static fileExtension(basename) {
 		const fileParts = basename.split(".");
 		if (fileParts <= 1)
 			return false;
 		return fileParts.pop();
-	},
+	}
 
 	/**
 	 * Create time string in format `1`d `1`h `1`m `1`s.
@@ -85,7 +85,7 @@ const Utils = {
 	 * formatTimeString(100500);
 	 * // "1d 3h 55m"
 	 */
-	formatTimeString(seconds) {
+	static formatTimeString(seconds) {
 		seconds = Math.floor(seconds);
 		let minutes = Math.floor(seconds / 60);
 		let hours = Math.floor(minutes / 60);
@@ -102,19 +102,19 @@ const Utils = {
 			result.unshift(`${days}d`);
 
 		return result.join(" ");
-	},
+	}
 
 	/**
 	 * Escape special markdown symbol "`" and wrapping this string into code block (with single quote).
 	 * @param {string} value Original string.
 	 * @return {string} Escaped string. If after escaping this string will empty then it returns empty string.
 	 */
-	escapeMarkdown(value) {
+	static escapeMarkdown(value) {
 		let escaped = value.replace(/`/g, "");
 		if (escaped.length === 0)
 			return "";
 		return `\`${escaped}\``;
 	}
-};
+}
 
 module.exports = Utils;
