+++ lib/classes/statistics/VoiceStateManager.js
@@ -1,6 +1,6 @@
-const { createModuleLog } = require("./Utils");
+const { createModuleLog } = require("../Utils");
 const log = createModuleLog("VoiceStateManager");
-const { VoiceStatisticsModel, GuildMemberModel } = require("./Database");
+const { VoiceStatisticsModel, GuildMemberModel } = require("../Database");
 
 /**
  * # Voice State Manager
