+++ lib/commands/basic/PingCommand.js
@@ -1,4 +1,4 @@
-const Bot = require("../../classes/Bot.js");
+const Bot = require("../../classes/CognitumClient.js");
 const BaseCommand = require("../../classes/base/BaseCommand.js");
 const CoreCategory = require("../../categories/CoreCategory.js");
 const DefaultEmbed = require("../../classes/embed/DefaultEmbed.js");
